feat(convention): update filenames to feat with project naming conventioin
