+++ src/onboarding/handlers/components_handler.ts
@@ -2,7 +2,7 @@ import * as fs from "fs";
 import * as path from 'path';
 import * as dotenv from "dotenv";
 import CustomComponent from "../components/CustomComponent";
-import { UtilsManager } from "../utils/UtilsManager";
+import { ClientManager } from "../utils/client_manager";
 
 export default async () => {
   dotenv.config();
@@ -27,6 +27,6 @@ export default async () => {
     if (file.includes('CustomComponent'))
       continue;
     const component: CustomComponent = new (await import(`../../${file}`)).default as CustomComponent;
-    UtilsManager.add_component(component);
+    ClientManager.add_component(component);
   }
 }
