+++ src/onboarding/index.ts
@@ -1,7 +1,7 @@
 import { Client, Events, GatewayIntentBits, Interaction } from "discord.js";
 import EventEmitter from "events";
 import * as dotenv from "dotenv";
-import { UtilsManager } from "./utils/UtilsManager";
+import { ClientManager } from "./utils/client_manager";
 
 dotenv.config();
 process.setMaxListeners(0);
@@ -11,11 +11,11 @@ const DISCORD_TOKEN = process.env.DISCORD_TOKEN;
 const DISCORD_ID = process.env.DISCORD_ID;
 
 const client = new Client({ intents: [GatewayIntentBits.Guilds, GatewayIntentBits.DirectMessages, GatewayIntentBits.GuildMessages, GatewayIntentBits.GuildMembers, GatewayIntentBits.GuildInvites] });
-UtilsManager.init(client, DISCORD_TOKEN, DISCORD_ID);
+ClientManager.init(client, DISCORD_TOKEN, DISCORD_ID);
 
 client.on(Events.InteractionCreate, async (interaction: Interaction) => {
-  if (interaction.isChatInputCommand()) { UtilsManager.get_command(interaction['commandName'])?.execute(interaction); } 
-  else { UtilsManager.get_component((interaction as any)['customId'])?.execute(interaction); }
+  if (interaction.isChatInputCommand()) { ClientManager.get_command(interaction['commandName'])?.execute(interaction); } 
+  else { ClientManager.get_component((interaction as any)['customId'])?.execute(interaction); }
 });
 
 client.login(DISCORD_TOKEN);
